Refactor CheckoutActivity and ElementsActivity
