+++ examples/basic-integration-jetpack-compose/app/src/main/java/com/deuna/compose_demo/MainActivity.kt
@@ -28,8 +28,8 @@ class MainActivity() : ComponentActivity() {
               // Initialize the view model with DeunaSDK configuration for sandbox environment
               homeViewModel = HomeViewModel(
                 deunaSDK = DeunaSDK(
-                  environment = Environment.SANDBOX,
-                  publicApiKey = "YOUR_PUBLIC_API_KEY"
+                  environment = Environment.STAGING,
+                  publicApiKey = "e40affdfbee57e43de41d1ce1451859bbe85626c1e87adaa93e538a6fb68488d09bb578f561122c1177e66ab1238563359acb70aa0b972ac8f44a52bceb7"
                 ),
               )
             )
