+++ src/main/AndroidManifest.xml
@@ -5,10 +5,10 @@
 
     <application>
         <activity
-            android:name=".element.DeunaElementActivity"
+            android:name=".web_views.CheckoutActivity"
             android:exported="false" />
         <activity
-            android:name=".checkout.DeunaActivity"
+            android:name=".web_views.ElementsActivity"
             android:exported="false" />
     </application>
