+++ src/main/java/com/deuna/maven/checkout/domain/CheckoutBridge.kt
@@ -4,7 +4,6 @@ import CheckoutResponse
 import android.content.Context
 import com.deuna.maven.*
 import com.deuna.maven.shared.*
-import com.deuna.maven.webviews.*
 import org.json.*
 
 class CheckoutBridge(
