+++ src/main/java/com/deuna/maven/element/domain/ElementsBridge.kt
@@ -4,7 +4,6 @@ import ElementsResponse
 import android.content.Context
 import com.deuna.maven.*
 import com.deuna.maven.shared.*
-import com.deuna.maven.webviews.*
 import org.json.*
 
  class ElementsBridge(
