+++ src/main/java/com/deuna/maven/web_views/ElementsActivity.kt
@@ -1,32 +1,43 @@
-package com.deuna.maven.webviews
+package com.deuna.maven.web_views
 
 import android.os.*
 import com.deuna.maven.element.domain.*
 import com.deuna.maven.shared.*
 
+/**
+ * This activity displays a web page loaded from a provided URL. It utilizes the ElementsBridge
+ * for communication between the WebView and the native Android code. It also interacts with the
+ * ElementsCallbacks interface to report success, errors, or cancellation.
+ */
 class ElementsActivity() : BaseWebViewActivity() {
 
     companion object {
         const val EXTRA_URL = "EXTRA_URL"
         private var callbacks: ElementsCallbacks? = null
 
+        /**
+         * Set the callbacks object to receive element events.
+         */
         fun setCallbacks(callbacks: ElementsCallbacks) {
             this.callbacks = callbacks
         }
     }
 
+    // Set of ElementsEvents indicating when to close the activity
     private lateinit var closeEvents: Set<ElementsEvent>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
+        // Extract the URL from the intent
         val url = intent.getStringExtra(EXTRA_URL)!!
 
+        // Parse close events from intent (if any)
         val closeEventAsStrings = intent.getStringArrayListExtra(EXTRA_CLOSE_EVENTS) ?: emptyList<String>()
         closeEvents = parseCloseEvents<ElementsEvent>(closeEventAsStrings)
 
+        // Load the provided URL
         loadUrl(url)
-
     }
 
     override fun getBridge(): WebViewBridge {
@@ -37,6 +48,8 @@ class ElementsActivity() : BaseWebViewActivity() {
         )
     }
 
+
+    // Notify callbacks about no internet connection
     override fun onNoInternet() {
         callbacks?.onError?.invoke(
             ElementsError(
@@ -50,6 +63,7 @@ class ElementsActivity() : BaseWebViewActivity() {
     }
 
     override fun onDestroy() {
+        // Notify callbacks about activity closure
         callbacks?.onClosed?.invoke()
         super.onDestroy()
     }
